Pull out some classes and functions for progress bars into progress.py.
This is the first code outside the main file, and means that test_b2_command_line.py
can no longer say "python b2/b2.py" to run it.  It now installs b2 into the current virtual
env and runs it there.  This has the benefit also testing the config in setup.py.
